requirements.txt
streamlit
pandas
numpy
ccxt
httpx
pyyaml
plotly
sqlalchemy
